fix(frontend/backend): i18n frontend rage
