+++ packages/backend/src/routes/users/index.ts
@@ -2,7 +2,6 @@ import Server from '../../server';
 
 import Route from '../route';
 
-import language from './language';
 import me from './me';
 import other from './other';
 
@@ -11,7 +10,6 @@ export default function generateRoute(server: Server): Route {
 
   route.use(me);
   route.use(other);
-  route.use(language);
 
   return route;
 }
