+++ packages/frontend/src/i18n/locales/ko/classroom.tsx
@@ -4,6 +4,7 @@ export default {
   start: '수업 시작하기',
   end: '수업 끝내기',
   changeVideo: '유튜브 영상 변경하기',
+  shareVideo: '유튜브 영상 공유하기',
   stopShare: '영상 공유 중지하기',
   startClass: '수업이 시작되었습니다.',
   endClass: '수업이 끝났습니다.',
@@ -15,4 +16,7 @@ export default {
   reset: '리셋',
   leave: '수업 나가기',
   remove: '수업 삭제하기',
+  speak: '말하기',
+  search: '유튜브 영상 검색',
+  share: '영상 공유',
 };
